LPS-75719 Move exported packages to new portal-workflow-api module.
