+++ portal-workflow-api/bnd.bnd
@@ -1,0 +1,9 @@
+Bundle-Name: Liferay Portal Workflow API
+Bundle-SymbolicName: com.liferay.portal.workflow.api
+Bundle-Version: 1.0.0
+Export-Package:\
+	com.liferay.portal.workflow.constants,\
+	com.liferay.portal.workflow.portlet.tab
+Liferay-Releng-Module-Group-Description:
+Liferay-Releng-Module-Group-Title: Workflow
+-dsannotations-options: inherit
