+++ portal-workflow-api/src/main/java/com/liferay/portal/workflow/portlet/tab/BaseWorkflowPortletTab.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.workflow.web.portlet.tab;
+package com.liferay.portal.workflow.portlet.tab;
 
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
