+++ portal-workflow-web/bnd.bnd
@@ -1,9 +1,6 @@
 Bundle-Name: Liferay Portal Workflow Web
 Bundle-SymbolicName: com.liferay.portal.workflow.web
 Bundle-Version: 1.0.0
-Export-Package:\
-	com.liferay.portal.workflow.web.constants,\
-	com.liferay.portal.workflow.web.portlet.tab
 Liferay-Releng-Module-Group-Description:
 Liferay-Releng-Module-Group-Title: Workflow
 Web-ContextPath: /portal-workflow-web
