+++ portal-workflow-web/build.gradle
@@ -20,6 +20,7 @@ dependencies {
 	provided group: "javax.servlet.jsp", name: "javax.servlet.jsp-api", version: "2.3.1"
 	provided group: "org.osgi", name: "org.osgi.core", version: "5.0.0"
 	provided group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
+	provided project(":apps:forms-and-workflow:portal-workflow:portal-workflow-api")
 
 	testCompile group: "junit", name: "junit", version: "4.12"
 }
