+++ portal-workflow-web/src/main/java/com/liferay/portal/workflow/web/internal/portlet/BaseWorkflowPortlet.java
@@ -20,8 +20,8 @@ import com.liferay.portal.kernel.portlet.bridges.mvc.MVCPortlet;
 import com.liferay.portal.kernel.servlet.SessionErrors;
 import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.workflow.WorkflowException;
-import com.liferay.portal.workflow.web.constants.WorkflowWebKeys;
-import com.liferay.portal.workflow.web.portlet.tab.WorkflowPortletTab;
+import com.liferay.portal.workflow.constants.WorkflowWebKeys;
+import com.liferay.portal.workflow.portlet.tab.WorkflowPortletTab;
 
 import java.io.IOException;
