+++ portal-workflow-web/src/main/java/com/liferay/portal/workflow/web/internal/portlet/tab/WorkflowInstancePortletTab.java
@@ -25,11 +25,11 @@ import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.portal.kernel.workflow.WorkflowInstance;
 import com.liferay.portal.kernel.workflow.WorkflowInstanceManagerUtil;
-import com.liferay.portal.workflow.web.constants.WorkflowWebKeys;
+import com.liferay.portal.workflow.constants.WorkflowWebKeys;
+import com.liferay.portal.workflow.portlet.tab.BaseWorkflowPortletTab;
+import com.liferay.portal.workflow.portlet.tab.WorkflowPortletTab;
 import com.liferay.portal.workflow.web.internal.configuration.WorkflowInstanceWebConfiguration;
 import com.liferay.portal.workflow.web.internal.request.prepocessor.WorkflowPreprocessorHelper;
-import com.liferay.portal.workflow.web.portlet.tab.BaseWorkflowPortletTab;
-import com.liferay.portal.workflow.web.portlet.tab.WorkflowPortletTab;
 
 import java.util.Map;
