+++ portal-workflow-web/src/main/resources/META-INF/resources/init.jsp
@@ -22,9 +22,9 @@
 taglib uri="http://liferay.com/tld/util" prefix="liferay-util" %>
 
 <%@ page import="com.liferay.portal.kernel.util.ParamUtil" %><%@
-page import="com.liferay.portal.workflow.web.constants.WorkflowWebKeys" %><%@
+page import="com.liferay.portal.workflow.constants.WorkflowWebKeys" %><%@
+page import="com.liferay.portal.workflow.portlet.tab.WorkflowPortletTab" %><%@
 page import="com.liferay.portal.workflow.web.internal.constants.WorkflowPortletKeys" %><%@
-page import="com.liferay.portal.workflow.web.portlet.tab.WorkflowPortletTab" %><%@
 page import="com.liferay.taglib.servlet.PipingServletResponse" %>
 
 <%@ page import="java.util.List" %>
